Зародыш облачного хранилища. Еще много работы. Класс Client тебует оптимизации Работает обрезанная Аутентификация с возвращением статуса Работает визуализация файл листа, как локальная так и серверная.

+++ common/src/main/java/com/geekstomach/cloud/common/protocol/Command.java
@@ -14,8 +14,8 @@ public final class Command {
     public static final short REGISTER = 9;
 //handel files
 public static final byte FILE_UPLOAD = 10;
-    public static final short FILE_DOWNLOAD = 11;
-    public static final short FILE_GET_LIST = 12;
+    public static final byte FILE_DOWNLOAD = 11;
+    public static final byte FILE_GET_LIST = 12;
     public static final short FILE_DELETE_FILE = 13;
 
 //handle cloud
